sim/requirements.txt
cocotb==1.9.2
cocotb-bus==0.2.1
cocotb-test==0.2.6
cocotbext-axi==0.1.24
find-libpython==0.4.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
pytest==8.4.1
